--- CropperSample/src/main/java/com/example/croppersample/MainActivity.java
@@ -1,281 +1,0 @@
-
-package com.example.croppersample;
-
-import android.app.Activity;
-import android.content.ComponentName;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
-import android.graphics.Bitmap;
-import android.net.Uri;
-import android.os.Bundle;
-import android.os.Parcelable;
-import android.provider.MediaStore;
-import android.view.View;
-import android.view.Window;
-import android.widget.AdapterView;
-import android.widget.Button;
-import android.widget.CompoundButton;
-import android.widget.CompoundButton.OnCheckedChangeListener;
-import android.widget.ImageView;
-import android.widget.SeekBar;
-import android.widget.SeekBar.OnSeekBarChangeListener;
-import android.widget.Spinner;
-import android.widget.TextView;
-import android.widget.ToggleButton;
-
-import com.edmodo.cropper.CropImageView;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
-
-public class MainActivity extends Activity {
-
-    // Static final constants
-    private static final int DEFAULT_ASPECT_RATIO_VALUES = 20;
-
-    private static final int ROTATE_NINETY_DEGREES = 90;
-
-    private static final String ASPECT_RATIO_X = "ASPECT_RATIO_X";
-
-    private static final String ASPECT_RATIO_Y = "ASPECT_RATIO_Y";
-
-    private static final int ON_TOUCH = 1;
-
-    // Instance variables
-    private int mAspectRatioX = DEFAULT_ASPECT_RATIO_VALUES;
-
-    private int mAspectRatioY = DEFAULT_ASPECT_RATIO_VALUES;
-
-    Bitmap croppedImage;
-
-    // Saves the state upon rotating the screen/restarting the activity
-    @Override
-    protected void onSaveInstanceState(@SuppressWarnings("NullableProblems") Bundle bundle) {
-        super.onSaveInstanceState(bundle);
-        bundle.putInt(ASPECT_RATIO_X, mAspectRatioX);
-        bundle.putInt(ASPECT_RATIO_Y, mAspectRatioY);
-    }
-
-    // Restores the state upon rotating the screen/restarting the activity
-    @Override
-    protected void onRestoreInstanceState(@SuppressWarnings("NullableProblems") Bundle bundle) {
-        super.onRestoreInstanceState(bundle);
-        mAspectRatioX = bundle.getInt(ASPECT_RATIO_X);
-        mAspectRatioY = bundle.getInt(ASPECT_RATIO_Y);
-    }
-
-    public void onCreate(Bundle savedInstanceState) {
-
-        super.onCreate(savedInstanceState);
-        this.requestWindowFeature(Window.FEATURE_NO_TITLE);
-        setContentView(R.layout.activity_main);
-
-        // Initialize components of the app
-        final CropImageView cropImageView = (CropImageView) findViewById(R.id.CropImageView);
-        final SeekBar aspectRatioXSeek = (SeekBar) findViewById(R.id.aspectRatioXSeek);
-        final SeekBar aspectRatioYSeek = (SeekBar) findViewById(R.id.aspectRatioYSeek);
-        Spinner showGuidelinesSpin = (Spinner) findViewById(R.id.showGuidelinesSpin);
-        final TextView aspectRatioNum = (TextView) findViewById(R.id.aspectRatioNum);
-
-        // Sets sliders to be disabled until fixedAspectRatio is set
-        aspectRatioXSeek.setEnabled(false);
-        aspectRatioYSeek.setEnabled(false);
-
-        // Set initial spinner value
-        showGuidelinesSpin.setSelection(ON_TOUCH);
-
-        //Sets the rotate button
-        final Button rotateButton = (Button) findViewById(R.id.Button_rotate);
-        rotateButton.setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View v) {
-                cropImageView.rotateImage(ROTATE_NINETY_DEGREES);
-            }
-        });
-
-        // Sets fixedAspectRatio
-        ((ToggleButton) findViewById(R.id.fixedAspectRatioToggle)).setOnCheckedChangeListener(new OnCheckedChangeListener() {
-
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                cropImageView.setFixedAspectRatio(isChecked);
-                if (isChecked) {
-                    aspectRatioXSeek.setEnabled(true);
-                    aspectRatioYSeek.setEnabled(true);
-                } else {
-                    aspectRatioXSeek.setEnabled(false);
-                    aspectRatioYSeek.setEnabled(false);
-                }
-            }
-        });
-
-        // Sets crop shape
-        ((ToggleButton) findViewById(R.id.cropShapeToggle)).setOnCheckedChangeListener(new OnCheckedChangeListener() {
-
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                cropImageView.setCropShape(isChecked ? CropImageView.CropShape.OVAL : CropImageView.CropShape.RECTANGLE);
-            }
-        });
-
-        // Sets initial aspect ratio to 10/10, for demonstration purposes
-        cropImageView.setAspectRatio(DEFAULT_ASPECT_RATIO_VALUES, DEFAULT_ASPECT_RATIO_VALUES);
-
-        aspectRatioXSeek.setOnSeekBarChangeListener(new OnSeekBarChangeListener() {
-            @Override
-            public void onProgressChanged(SeekBar aspectRatioXSeek, int progress, boolean fromUser) {
-                mAspectRatioX = progress;
-                cropImageView.setAspectRatio(progress, mAspectRatioY);
-                aspectRatioNum.setText("(" + mAspectRatioX + ", " + mAspectRatioY + ")");
-            }
-
-            @Override
-            public void onStartTrackingTouch(SeekBar seekBar) {
-            }
-
-            @Override
-            public void onStopTrackingTouch(SeekBar seekBar) {
-            }
-        });
-
-        aspectRatioYSeek.setOnSeekBarChangeListener(new OnSeekBarChangeListener() {
-            @Override
-            public void onProgressChanged(SeekBar aspectRatioYSeek, int progress, boolean fromUser) {
-                mAspectRatioY = progress;
-                cropImageView.setAspectRatio(mAspectRatioX, progress);
-                aspectRatioNum.setText("(" + mAspectRatioX + ", " + mAspectRatioY + ")");
-            }
-
-            @Override
-            public void onStartTrackingTouch(SeekBar seekBar) {
-            }
-
-            @Override
-            public void onStopTrackingTouch(SeekBar seekBar) {
-            }
-        });
-
-        // Sets up the Spinner
-        showGuidelinesSpin.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
-            public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
-                cropImageView.setGuidelines(i);
-            }
-
-            public void onNothingSelected(AdapterView<?> adapterView) {
-                return;
-            }
-        });
-
-        findViewById(R.id.Button_crop).setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View v) {
-                croppedImage = cropImageView.getCroppedImage();
-                ImageView croppedImageView = (ImageView) findViewById(R.id.croppedImageView);
-                croppedImageView.setImageBitmap(croppedImage);
-            }
-        });
-
-        findViewById(R.id.Button_load).setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View v) {
-                startActivityForResult(getPickImageChooserIntent(), 200);
-            }
-        });
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (resultCode == Activity.RESULT_OK) {
-            Uri imageUri = getPickImageResultUri(data);
-            ((CropImageView) findViewById(R.id.CropImageView)).setImageUri(imageUri);
-        }
-    }
-
-    /**
-     * Create a chooser intent to select the source to get image from.<br/>
-     * The source can be camera's (ACTION_IMAGE_CAPTURE) or gallery's (ACTION_GET_CONTENT).<br/>
-     * All possible sources are added to the intent chooser.
-     */
-    public Intent getPickImageChooserIntent() {
-
-        // Determine Uri of camera image to save.
-        Uri outputFileUri = getCaptureImageOutputUri();
-
-        List<Intent> allIntents = new ArrayList<>();
-        PackageManager packageManager = getPackageManager();
-
-        // collect all camera intents
-        Intent captureIntent = new Intent(android.provider.MediaStore.ACTION_IMAGE_CAPTURE);
-        List<ResolveInfo> listCam = packageManager.queryIntentActivities(captureIntent, 0);
-        for (ResolveInfo res : listCam) {
-            Intent intent = new Intent(captureIntent);
-            intent.setComponent(new ComponentName(res.activityInfo.packageName, res.activityInfo.name));
-            intent.setPackage(res.activityInfo.packageName);
-            if (outputFileUri != null) {
-                intent.putExtra(MediaStore.EXTRA_OUTPUT, outputFileUri);
-            }
-            allIntents.add(intent);
-        }
-
-        // collect all gallery intents
-        Intent galleryIntent = new Intent(Intent.ACTION_GET_CONTENT);
-        galleryIntent.setType("image/*");
-        List<ResolveInfo> listGallery = packageManager.queryIntentActivities(galleryIntent, 0);
-        for (ResolveInfo res : listGallery) {
-            Intent intent = new Intent(galleryIntent);
-            intent.setComponent(new ComponentName(res.activityInfo.packageName, res.activityInfo.name));
-            intent.setPackage(res.activityInfo.packageName);
-            allIntents.add(intent);
-        }
-
-        // the main intent is the last in the list (fucking android) so pickup the useless one
-        Intent mainIntent = allIntents.get(allIntents.size() - 1);
-        for (Intent intent : allIntents) {
-            if (intent.getComponent().getClassName().equals("com.android.documentsui.DocumentsActivity")) {
-                mainIntent = intent;
-                break;
-            }
-        }
-        allIntents.remove(mainIntent);
-
-        // Create a chooser from the main intent
-        Intent chooserIntent = Intent.createChooser(mainIntent, "Select source");
-
-        // Add all other intents
-        chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, allIntents.toArray(new Parcelable[allIntents.size()]));
-
-        return chooserIntent;
-    }
-
-    /**
-     * Get URI to image received from capture by camera.
-     */
-    private Uri getCaptureImageOutputUri() {
-        Uri outputFileUri = null;
-        File getImage = getExternalCacheDir();
-        if (getImage != null) {
-            outputFileUri = Uri.fromFile(new File(getImage.getPath(), "pickImageResult.jpeg"));
-        }
-        return outputFileUri;
-    }
-
-    /**
-     * Get the URI of the selected image from {@link #getPickImageChooserIntent()}.<br/>
-     * Will return the correct URI for camera and gallery image.
-     *
-     * @param data the returned data of the activity result
-     */
-    public Uri getPickImageResultUri(Intent data) {
-        boolean isCamera = true;
-        if (data != null) {
-            String action = data.getAction();
-            isCamera = action != null && action.equals(MediaStore.ACTION_IMAGE_CAPTURE);
-        }
-        return isCamera ? getCaptureImageOutputUri() : data.getData();
-    }
-}
