+++ cropper/src/main/res/layout/crop_image_view.xml
@@ -10,7 +10,7 @@
         tools:ignore="contentDescription"
         android:adjustViewBounds="true"/>
 
-    <com.edmodo.cropper.cropwindow.CropOverlayView
+    <com.theartofdev.edmodo.cropper.cropwindow.CropOverlayView
         android:id="@+id/CropOverlayView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
