+++ sample/src/main/res/drawable-xhdpi/butterfly.jpg
