+++ sample/src/main/res/drawable-xhdpi/violetsky.png
