+++ settings.gradle
@@ -1,2 +1,2 @@
 include 'cropper'
-include 'CropperSample'
+include 'sample'
